Use GcsStorageFactory in tests
